requirements.txt
django==4.2.7
django-extensions
django-filter
ipython

#flake8
flake8
flake8-builtins
flake8-print

django-crispy-forms
crispy-bootstrap4
Celery

names